pyproject.toml
[project]
name = "yuxu"
version = "0.0.1"
description = "Yuxu (玉虚) — a long-running agent creation and supervision framework"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "yuxu contributors"}]
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff",
]

[project.scripts]
yuxu = "yuxu.cli.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
yuxu = [
    "bundled/**/*.md",
    "bundled/**/*.yaml",
    "bundled/**/*.yml",
    "templates/**/*",
    "templates/**/.gitignore",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
